Updating iUTAH YODA examples
Not sure what happened when things were moved to the new repository.  I
think my newest versions got put in the "prototypes" folder, and some
older ones with "+AKA" on the end of them were left in the examples
folder. I moved the older files to the prototypes folder and put the
newest versions in the examples folder. The newest versions have
several updates to reflect changes made on the ODM2 teleconference this
week and replacement of blanks and empty strings with "NULL" values.