Added opening macro to stop calculation
I thought I'd had calculation turned off to the header and generator
sheets, but apparently it was still working on open until the generation
macro was run the first time.  I added an 'on open' macro which will
just turn those two sheets off so they only calculate when the macro is
called.