Work on CV update
Tried to make the CV update code work on both windows and Mac. The
modified code works on Windows, but still doesn’t work on Mac -
although it should. I suspect a file access issue in downloading the CV
files from the CV server related to “sandboxing” of Office applications
on OS X.