Updated prototype Specimens template
I think this one is good. I have questions for others on how to proceed
with data value level attributes (i.e. censor codes). This is an issue
that is equally important for the Time Series templates (i.e. I think
it is done incorrectly in those). Next step, generate a mockup YODA
file from this, to guide output.
Also, minor typo fix in Profiles doc.